ViTMusicToSequence/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.28.0
numpy==1.24.3
Pillow==9.4.0
opencv-python==4.5.5.64